Fix infinite fetch loop when calling useQuery with object literal
Before, when declaring parameters using e.g. an object literal and
passing them to useQuery, this would cause an infinite fetch loop
as useQuery would refetch every time the object identity of the
parameters changed, which happened on every render.

Now, the new parameters are checked for changes against the old
parameters using deep equality in useQuery, so refetching only
occurs if the values of the parameters or the query actually change.